Refac: implement OOP with es6 module
